refactor(analysis): extract BaseLanguageAnalyzer to remove duplication
