+++ analysis/src/main/kotlin/de/maibornwolff/dependacharta/pipeline/analysis/analyzers/BaseLanguageAnalyzer.kt
@@ -1,0 +1,49 @@
+package de.maibornwolff.dependacharta.pipeline.analysis.analyzers
+
+import de.maibornwolff.dependacharta.pipeline.analysis.model.*
+import de.maibornwolff.dependacharta.pipeline.shared.SupportedLanguage
+import de.maibornwolff.treesitter.excavationsite.api.Declaration
+import de.maibornwolff.treesitter.excavationsite.api.Language
+import de.maibornwolff.treesitter.excavationsite.api.TreeSitterDependencies
+
+abstract class BaseLanguageAnalyzer(
+    private val fileInfo: FileInfo,
+) : LanguageAnalyzer {
+    protected abstract val language: SupportedLanguage
+
+    override fun analyze(): FileReport {
+        val result = TreeSitterDependencies.analyze(fileInfo.content, Language.valueOf(language.name))
+        val imports = result.imports.map { it.toDependency() }
+        val dependencies = if (result.packagePath.isNotEmpty()) {
+            (imports + Dependency(path = Path(result.packagePath), isWildcard = true)).toSet()
+        } else {
+            imports.toSet()
+        }
+        val nodes = result.declarations.map { declaration ->
+            toNode(result.packagePath, dependencies, declaration)
+        }
+        return FileReport(nodes)
+    }
+
+    protected open fun buildPathWithName(
+        packagePath: List<String>,
+        declaration: Declaration,
+    ): Path {
+        return Path(packagePath + declaration.name)
+    }
+
+    private fun toNode(
+        packagePath: List<String>,
+        dependencies: Set<Dependency>,
+        declaration: Declaration,
+    ): Node {
+        return Node(
+            pathWithName = buildPathWithName(packagePath, declaration),
+            physicalPath = fileInfo.physicalPath,
+            language = language,
+            nodeType = declaration.type.toNodeType(),
+            dependencies = dependencies,
+            usedTypes = declaration.usedTypes.map { it.toType() }.toSet()
+        )
+    }
+}
